Layout Update - separated into php files for re-use
Created a generic _Layout.php file in the Layouts folder within the
website. It contains include lines for the .php files that house the top
navigation and side navigation bars. This way we only have to modify
things in one place (for navigation). Copy paste the _Layout.php and
update the relative links and it should be good to go.