Broad fixes/tweaks - Applying themes across pages, hooking up links.
I essentially created an addition nav index so that the files that
aren't one level deep (index, register, login) will have a link to one
another properly. The Styles folder also decided it didn't want to be
lower case, so i just toggled the references to it.

I also applied the themes to the login and register pages along with
moving them out of the login folder, and fixed up the relative link
paths within.